Wait for a WebSocket connection to close before reconnecting
As identified by @aguspdana, the original implementation attempted to
reconnect every so often, even if other connection attempts were still
in progress! This could lead to a situation where multiple WebSockets
were connected, which could explain the recent surge in connections we
saw *after* the incident where the server could not respond.