Perform a server handshake before treating the WebSocket as connected
There's an interesting difference when the playground is deployed to
production: there are now proxies between the web browser and the Axum
server! On the official deployment, we have:

```
Browser -> Cloudflare -> nginx -> Axum
```

The i32 deployment is the same but removes Cloudflare.

In these situations, the WebSocket connection will
effectively *always* succeed as the proxy accepts the
connection. However, once the proxy detects that the upstream server
is unavailable, it closes the WebSocket. On the browser side, that
manifests as the `open` and `close` events happening back-to-back.

The previous code treated the `open` event as a sign that the server
was active and listening and that attempting a reconnection was a
reasonable thing to do. This results in a lot of wasted
browser-initiated WebSocket connections when the proxy is available
but the upstream server is not.

The new code ensures that the server is truly available by sending a
message and waiting for the response. We also can double-dip by adding
a flag that might discourage people from also using the WebSocket
communication pathway themselves and expecting it to be a supported
API.