bincode is blocking
We spawn blocking tasks to process stdin/stdout from the coordinator
to the worker. We also remove some unneeded buffering (the `BufReader`
and `BufWriter` can take care of that for us) and some looping that
could never occur. We also remove the length prefix and allow bincode
to completely (de)serialize what it needs to.